Move new MathUtils to a better-named utility class
This class, Distributions, now gets tested in DistributionsTest, which
is a class that does not clash with our large internal product, which
for archaic reasons consumes some of our tests and gets indigestion
when a test name clashes with one of their test names.
