+++ atlasdb-commons/src/main/java/com/palantir/util/MathUtils.java
@@ -24,8 +24,9 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 
-import org.apache.commons.math3.special.Gamma;
-
+// WARNING: Do not create tests for this class! These methods are tested internally.
+// TODO(sberler): move these methods into the internal product where they are used.
+// If you are here to write a new util, please put it into a well-named file in the com.palantir.util.math package.
 public class MathUtils {
     // =========================
     // COMBINATORICS
@@ -247,19 +248,4 @@ public class MathUtils {
 
     private MathUtils() {/**/}
 
-    public static double calculateConfidenceThatDistributionIsNotUniform(List<Long> values) {
-        return 1.0 - calculateConfidenceThatDistributionIsUniform(values);
-    }
-
-    public static double calculateConfidenceThatDistributionIsUniform(List<Long> values) {
-        return Gamma.regularizedGammaQ((values.size() - 1.0) / 2,
-                chiSquareDistance(values) / 2);
-    }
-
-    private static double chiSquareDistance(List<Long> values) {
-        double mean = values.stream().mapToLong(x -> x).reduce(0L, Long::sum) / (double) values.size();
-        double distances =  values.stream().mapToDouble(Long::doubleValue)
-                .reduce(0.0, (acc, nextVal) -> acc + Math.pow(nextVal - mean, 2));
-        return distances / mean;
-    }
 }
