+++ atlasdb-commons/src/test/java/com/palantir/util/math/DistributionsTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.util;
+package com.palantir.util.math;
 
 import java.util.List;
 
@@ -24,33 +24,32 @@ import org.junit.Test;
 
 import com.google.common.collect.ImmutableList;
 
-public class MathUtilsTest {
-
+public class DistributionsTest {
     @Test
     public void universalDistributionConfidenceAlmostUniformDistributionTest() {
         List<Long> almostUniformDistribution = ImmutableList.of(199909L, 200065L, 199907L, 200070L, 199949L);
-        Assertions.assertThat(MathUtils.calculateConfidenceThatDistributionIsNotUniform(almostUniformDistribution))
+        Assertions.assertThat(Distributions.confidenceThatDistributionIsNotUniform(almostUniformDistribution))
                 .isCloseTo(0.0, Offset.offset(0.01));
     }
 
     @Test
     public void universalDistributionConfidenceRelativelyUniformDistributionTest() {
         List<Long> relativelyUniformDistribution = ImmutableList.of(199809L, 200665L, 199607L, 200270L, 199649L);
-        Assertions.assertThat(MathUtils.calculateConfidenceThatDistributionIsNotUniform(relativelyUniformDistribution))
+        Assertions.assertThat(Distributions.confidenceThatDistributionIsNotUniform(relativelyUniformDistribution))
                 .isLessThan(0.7);
     }
 
     @Test
     public void universalDistributionConfidenceSomeWhatSkewedDistributionTest() {
         List<Long> someWhatSkewedDistribution = ImmutableList.of(199609L, 200965L, 199607L, 200270L, 199649L);
-        Assertions.assertThat(MathUtils.calculateConfidenceThatDistributionIsNotUniform(someWhatSkewedDistribution))
+        Assertions.assertThat(Distributions.confidenceThatDistributionIsNotUniform(someWhatSkewedDistribution))
                 .isLessThan(0.88);
     }
 
     @Test
     public void universalDistributionConfidenceCompletelySkewedDistributionTest() {
         List<Long> completelySkewedDistribution = ImmutableList.of(522573L, 244456L, 139979L, 71531L, 21461L);
-        Assertions.assertThat(MathUtils.calculateConfidenceThatDistributionIsNotUniform(completelySkewedDistribution))
+        Assertions.assertThat(Distributions.confidenceThatDistributionIsNotUniform(completelySkewedDistribution))
                 .isCloseTo(1.0, Offset.offset(0.000001));
     }
 }
